+++ Vault/test/utopia/vault/test/database/access/item/versioned/FilterByVersionedTestItem.scala
@@ -1,0 +1,15 @@
+package utopia.vault.test.database.access.item.versioned
+
+import utopia.vault.nosql.view.{FilterableView, FilterableViewWrapper}
+
+/**
+  * An interface which provides versioned test item -based filtering for other types of access 
+  * points.
+  * @param wrapped Wrapped access point. Expected to include versioned_test_item.
+  * @tparam A Type of the wrapped access class
+  * @author Mikko Hilpinen
+  * @since 04.07.2025, v1.22
+  */
+case class FilterByVersionedTestItem[+A <: FilterableView[A]](wrapped: A) 
+	extends FilterTestItems[A] with FilterableViewWrapper[A]
+