+++ Vault/test/utopia/vault/test/model/factory/item/VersionedTestItemFactory.scala
@@ -1,0 +1,34 @@
+package utopia.vault.test.model.factory.item
+
+import java.time.Instant
+
+/**
+  * Common trait for versioned test item-related factories which allow construction with 
+  * individual properties
+  * @tparam A Type of constructed instances
+  * @author Mikko Hilpinen
+  * @since 04.07.2025, v1.22
+  */
+trait VersionedTestItemFactory[+A]
+{
+	// ABSTRACT	--------------------
+	
+	/**
+	  * @param created New created to assign
+	  * @return Copy of this item with the specified created
+	  */
+	def withCreated(created: Instant): A
+	
+	/**
+	  * @param deprecatedAfter New deprecated after to assign
+	  * @return Copy of this item with the specified deprecated after
+	  */
+	def withDeprecatedAfter(deprecatedAfter: Instant): A
+	
+	/**
+	  * @param name New name to assign
+	  * @return Copy of this item with the specified name
+	  */
+	def withName(name: String): A
+}
+